refactor(types): поля открытых объектов данных вынесены из инференсера
Инференсер держал в себе знание о конкретных конструкциях 1С: что значит
`Новый Структура("К1, К2")`, `Новый ОписаниеТипов("Число")`, `Х.Вставить("Имя", Значение)`
и `Х.Колонки.Добавить("Имя", Тип)`. Это не вывод типов, а база знаний о платформе, вшитая
прямо в диспетчер по узлам выражения.

Вынесено в `OpenDataObjectInference` — компонент про объекты, у которых состав членов
складывается в коде, а не объявлен заранее: Структура, ФиксированнаяСтруктура,
Соответствие, ФиксированноеСоответствие, ТаблицаЗначений. Он знает ровно два источника
таких полей — конструктор и оператор-мутатор — и больше ничего.

Про вывод типов компонент не знает: типы значений отдаёт колбэк, как это уже сделано у
расчёта по потоку. Обратной зависимости на инференсер нет.

Заодно нашёлся дом для чтения строкового литерала: его брал из инференсера и
`TableCollectionInference` — ради этого метод был видимым в пакете. Теперь оба берут его
у компонента.

Инференсер: 1650 строк на develop → 1527.
